ResumeExtractor/requirements.txt
pymupdf
google-genai
pydantic
pytest
python-dotenv
